+++ PhotoOverlay/Presentation/PhotoList/ViewModel/PhotoItem.swift
@@ -1,0 +1,12 @@
+//
+//  PhotoItem.swift
+//  PhotoOverlay
+//
+//  Created by 황제하 on 2022/07/05.
+//
+
+import UIKit
+
+struct PhotoItem: Hashable {
+    let photo: UIImage
+}
